+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.7.1 - Unreleased
 
+### Fixed:
+* Fixed issue where imdelay would sometimes not respect antialias and interpolation
 
 ## Version 0.7.0 - Released 2023-08-15